safety-monitor/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pyzmq==26.2.0
PyYAML==6.0.2

# Pi / sensors
smbus2==0.5.0
RPi.GPIO==0.7.1

# Optional for mic noise (USB mic)
sounddevice==0.5.1
numpy==2.2.2
